change input type to long

+++ simple-counter/src/main/java/com/davidodari/simplecounter/ICurrentValueListener.kt
@@ -1,5 +1,5 @@
 package com.davidodari.simplecounter
 
 interface ICurrentValueListener {
-    fun onValueChanged(value: Int)
+    fun onValueChanged(value: Long)
 }
